rename files to match convention
