+++ TypeScript/app/update-strategies/default-strategy.ts
@@ -1,8 +1,8 @@
-import Item from "./../models/Item";
-import IUpdateStrategy from "./IUpdateStrategy";
+import Item from "../models/Item";
+import UpdateStrategy from "./update-strategy";
 
 
-export default class DefaultStrategy implements IUpdateStrategy {    
+export default class DefaultStrategy implements UpdateStrategy {    
     constructor(private qualityFactor : number) {}
 
     updateItem(item: Item) : Item {
