+++ TypeScript/app/update-strategies/no-update-strategy.ts
@@ -1,7 +1,7 @@
 import Item from "@/models/Item";
-import IUpdateStrategy from "./IUpdateStrategy";
+import UpdateStrategy from "./update-strategy";
 
-export default class NoUpdateStrategy implements IUpdateStrategy {
+export default class NoUpdateStrategy implements UpdateStrategy {
     updateItem(item: Item) : Item {        
         return new Item(item.name, item.sellIn, item.quality);
     }
