+++ TypeScript/app/update-strategies/update-strategy.ts
@@ -1,0 +1,6 @@
+import Item from "../models/Item";
+
+
+export default interface UpdateStrategy {
+    updateItem(item: Item) : Item;
+}
