--- TypeScript/app/updateStrategies/IUpdateStrategy.ts
@@ -1,6 +1,0 @@
-import Item from "./../models/Item";
-
-
-export default interface IUpdateStrategy {
-    updateItem(item: Item) : Item;
-}
